# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architecture SaaS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: Get all clients
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /projects/{id}/feedback:
    get:
      operationId: getProjectFeedback
      tags: [feedback]
      summary: Get client feedback for a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project feedback list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientFeedback"
    post:
      operationId: createProjectFeedback
      tags: [feedback]
      summary: Add client feedback for a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackBody"
      responses:
        "201":
          description: Feedback created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientFeedback"
  /projects/{id}/estimates:
    get:
      operationId: getProjectEstimates
      tags: [estimates]
      summary: Get estimates (BOQ) for a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project estimates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimatesResponse"
    post:
      operationId: createProjectEstimate
      tags: [estimates]
      summary: Add an estimate item to a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateBody"
      responses:
        "201":
          description: Estimate item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateItem"
  /estimates/{estimateId}:
    put:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate item
      security:
        - bearerAuth: []
      parameters:
        - name: estimateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateBody"
      responses:
        "200":
          description: Estimate updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateItem"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate item
      security:
        - bearerAuth: []
      parameters:
        - name: estimateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimate deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
...
    ClientFeedback:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        stageId:
          type: integer
          nullable: true
        stageName:
          type: string
          nullable: true
        feedbackText:
          type: string
        feedbackType:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - feedbackText
        - feedbackType
    CreateFeedbackBody:
      type: object
      properties:
        stageId:
          type: integer
          nullable: true
        feedbackText:
          type: string
        feedbackType:
          type: string
      required:
        - feedbackText
        - feedbackType
    EstimateItem:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        phaseName:
          type: string
        itemName:
          type: string
        quantity:
          type: number
        unit:
          type: string
          nullable: true
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - phaseName
        - itemName
        - quantity
        - unitPrice
        - totalPrice
    EstimatesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EstimateItem"
        totalCost:
          type: number
      required:
        - items
        - totalCost
    CreateEstimateBody:
      type: object
      properties:
        phaseName:
          type: string
        itemName:
          type: string
        quantity:
          type: number
        unit:
          type: string
          nullable: true
        unitPrice:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - phaseName
        - itemName
        - quantity
        - unitPrice
...